Added timeout for client connections

+++ src/Utils.cpp
@@ -194,7 +194,7 @@ namespace Utils
 	//finds the extension of a file in path
 	std::string get_file_extension(const std::string& file_path)
 	{
-		int pos = file_path.rfind(".");
+		size_t pos = file_path.rfind(".");
 		if (pos != std::string::npos && pos < file_path.size() - 1)
 		{
 			return file_path.substr(pos);
